requirements.txt
python-telegram-bot>=21.0,<22.0
telethon>=1.36,<2.0
sqlalchemy>=2.0,<3.0
apscheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
pytz>=2024.1
